Blake scheduler
